Cambios en la gui, ya funcionan todas las ordenes desde los botones y demás

+++ src/sistema/controladores/ListenerFrenar.java
@@ -48,8 +48,8 @@ public class ListenerFrenar extends ListenerOrdenes implements ActionListener {
 		}
 		else {
 			
-			Integer cantidad = panel.getCantidadFrenado();
-			Integer tiempo = panel.getTiempoFrenado();
+			Double cantidad = panel.getCantidadFrenado();
+			Double tiempo = panel.getTiempoFrenado();
 			
 			comando = "ciclista " + idciclista + " frena " + cantidad + " en " + tiempo;
 		}
